Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletnode-constructor-options.https-expected.txt

CONSOLE MESSAGE: line 23: TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [setup]  
